requirements.txt
Django>=4.0
djangorestframework
django-filter
Faker
drf-yasg
